Separate shards from execution
